# Simplifies the simulator design to provide better performance