--- a/hyper-lib/src/lib.rs
+++ b/hyper-lib/src/lib.rs
@@ -11,13 +11,3 @@ pub type TxId = u32;
 
 pub const MAX_OUTBOUND_CONNECTIONS: usize = 8;
 static SECS_TO_NANOS: u64 = 1_000_000_000;
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use rand::{self, RngCore};
-
-    pub(crate) fn get_random_txid() -> TxId {
-        rand::thread_rng().next_u32()
-    }
-}