--- a/hyper-lib/src/txreconciliation.rs
+++ b/hyper-lib/src/txreconciliation.rs
@@ -1,4 +1,4 @@
-use std::collections::HashSet;
+use hashbrown::HashSet;
 
 use itertools::Itertools;