This PR simplifies the original simulator design in the following ways to allow simulations to run much faster:

- Computes fanout targets just once per transaction, instead of once per peer returning the exact same ordering
- Uses `hashbrown::Hash{map, set}`, which are more performant than the ones on the standard library at the cost of having less DoS resistance, but that is something that should not concern our use case*
- Redesigns the simulator to simulate **a single transaction**. This is the way the simulator was being used anyway, but the data structures and messages supported sending multiple ones. This meant that the events handled by the event queue had to pass around vectors of a single transaction, which was highly inefficient (specially when the underlying `BinaryHeap` had to re-order elements).

\* This may be even more simplified. I think I can get rid of all (or most) maps and replace them with vectors, which should be even more performant

Current results:

| branch  |  fanout sim time | erlay sim time |
| ------------- | ------------- |  ------------- |
| master  | 2.3s | ~8.5s |
|this-branch |~0.8s | ~2.9s|