@sr-gi 2024-11-25
> \* This may be even more simplified. I think I can get rid of all (or most) maps and replace them with vectors, which should be even more performant

I'm going to merge this as is, and potentially improve it down the line if it is worth spending more time on this